Fixed?
Started putting parts of the editor into place. Stat evolvers are a
thing taht can be editied along with name
