+++ Assets/Database/GGJobDatabase.asset.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 7546e8395443f3d47bfe377c52cd6251
-timeCreated: 1473354755
+guid: 70af2a513a57e2f44be1426f8ba4cbde
+timeCreated: 1473395510
 licenseType: Free
 NativeFormatImporter:
   userData:
