+++ Assets/Database/GGQualityDatabase.asset
@@ -11,4 +11,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 7830d77cf4822a648b0e03b2af7d36ac, type: 3}
   m_Name: GGQualityDatabase
   m_EditorClassIdentifier: 
-  database: []
+  database:
+  - _name: Common
+    _icon: {fileID: 0}
+  - _name: Uncommon
+    _icon: {fileID: 0}
