+++ Assets/Editor/JobSystem/JobSystemDatabaseEditor.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 85db0cfa6245dc24ba2db7d1e12dbefb
-timeCreated: 1473317695
+guid: 96dd77e0c8f0861418e238bc2920a540
+timeCreated: 1473395615
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
